Added main function
